requirements.txt
streamlit
langchain
langchain-groq
python-dotenv
pathlib
ipykernel
Ollama
langchain-community
streamlit 
faiss-cpu
pypdf2
pypdf
bs4
pydantic
openai
langchain-google-genai 
google-generativeai
faster-whisper
